requirements.txt
asgiref==3.3.4
Django==2.1.4
django-cors-headers==2.4.0
importlib-metadata==4.5.0
PyJWT==2.1.0
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
